Draft OpenAI converters
Draft converters

Update tests

Update tests

Put output format converter tests in own file

Remove unused imports
