+++ genai-perf/genai_perf/inputs/inputs.py
@@ -24,19 +24,14 @@ from genai_perf.exceptions import GenAIPerfException
 from genai_perf.inputs.input_constants import (
     DEFAULT_BATCH_SIZE,
     DEFAULT_INPUT_DATA_JSON,
-    DEFAULT_OUTPUT_TOKENS_MEAN,
-    DEFAULT_TENSORRTLLM_MAX_TOKENS,
-    EMPTY_JSON_IN_OPENAI_PA_FORMAT,
-    EMPTY_JSON_IN_TENSORRTLLM_PA_FORMAT,
-    EMPTY_JSON_IN_VLLM_PA_FORMAT,
     MINIMUM_LENGTH,
     MINIMUM_STARTING_INDEX,
-    ModelSelectionStrategy,
     OutputFormat,
     PromptSource,
     dataset_url_map,
 )
 from genai_perf.inputs.inputs_config import InputsConfig
+from genai_perf.inputs.output_format_converter import OutputFormatConverterFactory
 from genai_perf.inputs.synthetic_image_generator import (
     ImageFormat,
     SyntheticImageGenerator,
@@ -419,685 +414,13 @@ class Inputs:
         return payload
 
     def _convert_generic_json_to_output_format(self, generic_dataset) -> Dict:
-        if (
-            self.config.output_format == OutputFormat.OPENAI_CHAT_COMPLETIONS
-            or self.config.output_format == OutputFormat.OPENAI_VISION
-            or self.config.output_format == OutputFormat.IMAGE_RETRIEVAL
-        ):
-            output_json = self._convert_generic_json_to_openai_chat_completions_format(
-                generic_dataset,
-            )
-        elif self.config.output_format == OutputFormat.OPENAI_COMPLETIONS:
-            output_json = self._convert_generic_json_to_openai_completions_format(
-                generic_dataset,
-            )
-        elif self.config.output_format == OutputFormat.OPENAI_EMBEDDINGS:
-            output_json = self._convert_generic_json_to_openai_embeddings_format(
-                generic_dataset
-            )
-        elif self.config.output_format == OutputFormat.RANKINGS:
-            output_json = self._convert_generic_json_to_rankings_format(generic_dataset)
-        elif self.config.output_format == OutputFormat.VLLM:
-            output_json = self._convert_generic_json_to_vllm_format(generic_dataset)
-        elif self.config.output_format == OutputFormat.TENSORRTLLM:
-            output_json = self._convert_generic_json_to_trtllm_format(generic_dataset)
-        elif self.config.output_format == OutputFormat.TENSORRTLLM_ENGINE:
-            output_json = self._convert_generic_json_to_trtllm_engine_format(
-                generic_dataset
-            )
-        else:
-            raise GenAIPerfException(
-                f"Output format {self.config.output_format} is not currently supported"
-            )
-
-        return output_json
-
-    def _convert_generic_json_to_openai_chat_completions_format(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        # TODO (TMA-1757): Implement a way to select a role for `text_input`
-        (
-            system_role_headers,
-            user_role_headers,
-            _,
-        ) = self._determine_json_feature_roles(dataset_json)
-        pa_json = self._populate_openai_chat_completions_output_json(
-            dataset_json,
-            system_role_headers,
-            user_role_headers,
-        )
-
-        return pa_json
-
-    def _convert_generic_json_to_openai_completions_format(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        (
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        ) = self._determine_json_feature_roles(dataset_json)
-        pa_json = self._populate_openai_completions_output_json(
-            dataset_json,
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        )
-
-        return pa_json
-
-    def _convert_generic_json_to_openai_embeddings_format(
-        self,
-        generic_dataset: Dict,
-    ) -> Dict[str, Any]:
-        pa_json: Dict[str, Any] = {"data": []}
-
-        for index, entry in enumerate(generic_dataset["rows"]):
-            iter_model_name = self._select_model_name(index)
-            payload = entry.get("payload", {})
-            input_values = payload.get("input")
-
-            if input_values is None:
-                raise ValueError("Missing required fields 'input' in dataset entry")
-            if not isinstance(input_values, list):
-                raise ValueError(
-                    f"Required field 'input' must be a list (actual: {type(input_values)})"
-                )
-
-            payload = {
-                "input": input_values,
-                "model": iter_model_name,
-            }
-
-            for key, value in self.config.extra_inputs.items():
-                payload[key] = value
-
-            pa_json["data"].append({"payload": [payload]})
-
-        return pa_json
-
-    def _contains_rankings_tei(self) -> bool:
-        """
-        Check if user specified that they are using the Hugging Face
-        Text Embeddings Interface for ranking models
-        """
-        if (
-            self.config.extra_inputs
-            and self.config.extra_inputs.get("rankings") == "tei"
-        ):
-            return True
-        return False
-
-    def _convert_generic_json_to_rankings_format(
-        self,
-        generic_dataset: Dict,
-    ) -> Dict[str, Any]:
-        pa_json: Dict[str, Any] = {"data": []}
-        use_tei_format = self._contains_rankings_tei()
-
-        for index, entry in enumerate(generic_dataset["rows"]):
-            iter_model_name = self._select_model_name(index)
-
-            payload = entry.get("payload", {})
-            query_values = payload.get("query")
-
-            if use_tei_format:
-                passage_values = payload.get("passages", [])
-                passage_values = [item.get("text", "") for item in passage_values]
-            else:
-                passage_values = payload.get("passages")
-
-            if query_values is None:
-                raise ValueError("Missing required fields 'query' in dataset entry")
-            if passage_values is None:
-                raise ValueError(
-                    f"Missing required fields '{'texts' if use_tei_format else 'passages'}' in dataset entry"
-                )
-            if not isinstance(passage_values, list):
-                raise ValueError(
-                    f"Required field '{'texts' if use_tei_format else 'passages'}' must be a list (actual: {type(passage_values)})"
-                )
-
-            if use_tei_format:
-                payload = {"query": query_values["text"], "texts": passage_values}
-            else:
-                payload = {
-                    "query": query_values,
-                    "passages": passage_values,
-                    "model": iter_model_name,
-                }
-
-            for key, value in self.config.extra_inputs.items():
-                if not (key == "rankings" and value == "tei"):
-                    payload[key] = value
-
-            pa_json["data"].append({"payload": [payload]})
-
-        return pa_json
-
-    def _convert_generic_json_to_vllm_format(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        (
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        ) = self._determine_json_feature_roles(dataset_json)
-
-        pa_json = self._populate_vllm_output_json(
-            dataset_json,
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        )
-
-        return pa_json
-
-    def _convert_generic_json_to_trtllm_format(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        (
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        ) = self._determine_json_feature_roles(dataset_json)
-
-        pa_json = self._populate_trtllm_output_json(
-            dataset_json,
-            system_role_headers,
-            user_role_headers,
-            text_input_headers,
-        )
-
-        return pa_json
-
-    def _convert_generic_json_to_trtllm_engine_format(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        pa_json = self._populate_trtllm_engine_output_json(
-            dataset_json,
-        )
-        return pa_json
+        converter = OutputFormatConverterFactory.create(self.config.output_format)
+        return converter.convert(generic_dataset, self.config)
 
     def _write_json_to_file(self, json_in_pa_format: Dict) -> None:
         filename = self.config.output_dir / DEFAULT_INPUT_DATA_JSON
         with open(str(filename), "w") as f:
             f.write(json.dumps(json_in_pa_format, indent=2))
-
-    def _determine_json_feature_roles(
-        self, dataset_json: Dict
-    ) -> Tuple[List[str], List[str], List[str]]:
-        SYSTEM_ROLE_LIST = ["system_prompt"]
-        USER_ROLE_LIST = ["question", "article"]
-        TEXT_INPUT_LIST = ["text_input"]
-
-        system_role_headers: List[str] = []
-        user_role_headers: List[str] = []
-        text_input_headers: List[str] = []
-
-        if "features" in dataset_json.keys():
-            for feature in dataset_json["features"]:
-                if feature in SYSTEM_ROLE_LIST:
-                    system_role_headers.append(feature)
-                if feature in USER_ROLE_LIST:
-                    user_role_headers.append(feature)
-                if feature in TEXT_INPUT_LIST:
-                    user_role_headers.append(feature)
-
-        assert (
-            system_role_headers is not None
-            or user_role_headers is not None
-            or text_input_headers is not None
-        )
-
-        return system_role_headers, user_role_headers, text_input_headers
-
-    def _select_model_name(self, index):
-        if self.config.model_selection_strategy == ModelSelectionStrategy.ROUND_ROBIN:
-            return self.config.model_name[index % len(self.config.model_name)]
-        elif self.config.model_selection_strategy == ModelSelectionStrategy.RANDOM:
-            return random.choice(self.config.model_name)
-        else:
-            raise GenAIPerfException(
-                f"Model selection strategy '{self.config.model_selection_strategy}' is unsupported"
-            )
-
-    def _populate_openai_chat_completions_output_json(
-        self,
-        dataset_json: Dict,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-    ) -> Dict:
-        pa_json = self._create_empty_openai_pa_json()
-
-        for index, entry in enumerate(dataset_json["rows"]):
-            iter_model_name = self._select_model_name(index)
-            openai_json: Dict = {"payload": [{"messages": []}]}
-
-            # Check if the entry is a list (batched entries) or a single entry
-            if isinstance(entry, list):
-                for item in entry:
-                    self._process_row_content(
-                        item, system_role_headers, user_role_headers, openai_json
-                    )
-            elif isinstance(entry, dict):
-                self._process_row_content(
-                    entry, system_role_headers, user_role_headers, openai_json
-                )
-            else:
-                raise GenAIPerfException(f"Unexpected data type in rows: {type(entry)}")
-
-            self._add_optional_tags_to_openai_json(
-                openai_json,
-                iter_model_name,
-            )
-            pa_json["data"].append(openai_json)
-
-        return pa_json
-
-    def _process_row_content(
-        self,
-        entry: Dict,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        openai_json: Dict,
-    ) -> None:
-        if "image" in entry:
-            contents = self._extract_chat_contents(entry)
-            if openai_json["payload"][0]["messages"]:
-                openai_json["payload"][0]["messages"][0]["content"].extend(contents)
-            else:
-                openai_json["payload"][0]["messages"].append(
-                    {"role": "user", "content": contents}
-                )
-        else:
-            for header, content in entry.items():
-                message = self._create_new_openai_chat_completions_message(
-                    header, system_role_headers, user_role_headers, content
-                )
-                self._add_message_to_json(openai_json, message)
-
-    def _extract_chat_contents(self, entry: Dict) -> List[Dict]:
-        contents: List = []
-        if isinstance(entry, list):
-            for item in entry:
-                for content_type, content in item.items():
-                    self._add_content(contents, content_type, content)
-        else:
-            for content_type, content in entry.items():
-                self._add_content(contents, content_type, content)
-        return contents
-
-    def _add_content(self, contents: List[Dict], content_type: str, content: str):
-        if content_type == "text_input":
-            contents.append({"type": "text", "text": content})
-        elif content_type == "image":
-            contents.append({"type": "image_url", "image_url": {"url": content}})
-        else:
-            raise GenAIPerfException(
-                "Failed to construct OpenAI chat completions message "
-                f"contents. Unknown content type: '{content_type}'."
-            )
-
-    def _populate_openai_completions_output_json(
-        self,
-        dataset_json: Dict,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        text_input_headers: List[str],
-    ) -> Dict:
-        pa_json = self._create_empty_openai_pa_json()
-
-        for index, entry in enumerate(dataset_json["rows"]):
-            iter_model_name = self._select_model_name(index)
-            pa_json["data"].append({"payload": []})
-            pa_json["data"][index]["payload"].append({"prompt": ""})
-
-            for header, content in entry.items():
-                new_prompt = self._create_new_prompt(
-                    header,
-                    system_role_headers,
-                    user_role_headers,
-                    text_input_headers,
-                    content,
-                )
-
-                pa_json = self._add_new_prompt_to_json(pa_json, index, new_prompt)
-
-            self._add_optional_tags_to_openai_json(
-                pa_json["data"][index],
-                iter_model_name,
-            )
-
-        return pa_json
-
-    def _populate_vllm_output_json(
-        self,
-        dataset_json: Dict,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        text_input_headers: List[str],
-    ) -> Dict:
-        pa_json = self._create_empty_vllm_pa_json()
-
-        for index, entry in enumerate(dataset_json["rows"]):
-            iter_model_name = self._select_model_name(index)
-            pa_json["data"].append({"text_input": [""]})
-
-            for header, content in entry.items():
-                new_text_input = self._create_new_text_input(
-                    header,
-                    system_role_headers,
-                    user_role_headers,
-                    text_input_headers,
-                    content,
-                )
-
-                pa_json = self._add_new_text_input_to_json(
-                    pa_json, index, new_text_input
-                )
-
-            pa_json = self._add_optional_tags_to_vllm_json(
-                pa_json,
-                index,
-                iter_model_name,
-            )
-
-        return pa_json
-
-    def _populate_trtllm_output_json(
-        self,
-        dataset_json: Dict,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        text_input_headers: List[str],
-    ) -> Dict:
-        pa_json = self._create_empty_trtllm_pa_json()
-        default_max_tokens = (
-            "max_tokens" not in self.config.extra_inputs
-            or self.config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN
-        )
-
-        for index, entry in enumerate(dataset_json["rows"]):
-            iter_model_name = self._select_model_name(index)
-            pa_json["data"].append({"text_input": [""]})
-
-            for header, content in entry.items():
-                new_text_input = self._create_new_text_input(
-                    header,
-                    system_role_headers,
-                    user_role_headers,
-                    text_input_headers,
-                    content,
-                )
-
-                pa_json = self._add_new_text_input_to_json(
-                    pa_json, index, new_text_input
-                )
-
-            pa_json = self._add_required_tags_to_trtllm_json(
-                pa_json, index, default_max_tokens
-            )
-            pa_json = self._add_optional_tags_to_trtllm_json(
-                pa_json,
-                index,
-                iter_model_name,
-            )
-
-        return pa_json
-
-    def _populate_trtllm_engine_output_json(
-        self,
-        dataset_json: Dict,
-    ) -> Dict:
-        pa_json = self._create_empty_trtllm_pa_json()
-
-        for index, entry in enumerate(dataset_json["rows"]):
-            token_ids = self.config.tokenizer.encode(entry["text_input"])
-            pa_json["data"].append(
-                {
-                    "input_ids": {
-                        "content": token_ids,
-                        "shape": [len(token_ids)],
-                    },
-                    "input_lengths": [len(token_ids)],
-                    "request_output_len": [DEFAULT_TENSORRTLLM_MAX_TOKENS],
-                }
-            )
-
-            pa_json = self._add_optional_tags_to_trtllm_engine_json(
-                pa_json,
-                index,
-            )
-        return pa_json
-
-    def _create_empty_openai_pa_json(self) -> Dict:
-        empty_pa_json = deepcopy(EMPTY_JSON_IN_OPENAI_PA_FORMAT)
-
-        return empty_pa_json
-
-    def _create_empty_vllm_pa_json(self) -> Dict:
-        empty_pa_json = deepcopy(EMPTY_JSON_IN_VLLM_PA_FORMAT)
-
-        return empty_pa_json
-
-    def _create_empty_trtllm_pa_json(self) -> Dict:
-        empty_pa_json = deepcopy(EMPTY_JSON_IN_TENSORRTLLM_PA_FORMAT)
-
-        return empty_pa_json
-
-    def _create_new_openai_chat_completions_message(
-        self,
-        header: str,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        content: str,
-    ) -> Optional[Dict]:
-        # Do not add messages with blank content
-        if not content:
-            return {}
-
-        if header in system_role_headers:
-            message = {
-                "role": "system",
-                "content": content,
-            }
-        elif header in user_role_headers:
-            message = {
-                "role": "user",
-                "content": content,
-            }
-        else:
-            message = {}
-
-        return message
-
-    def _create_new_prompt(
-        self,
-        header: str,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        text_input_headers: List[str],
-        content: str,
-    ) -> str:
-        new_prompt = ""
-
-        if (
-            header in system_role_headers
-            or header in user_role_headers
-            or header in text_input_headers
-        ):
-            new_prompt = content
-
-        return new_prompt
-
-    def _create_new_text_input(
-        self,
-        header: str,
-        system_role_headers: List[str],
-        user_role_headers: List[str],
-        text_input_headers: List[str],
-        content: str,
-    ) -> str:
-        new_text_input = ""
-
-        if (
-            header in system_role_headers
-            or header in user_role_headers
-            or header in text_input_headers
-        ):
-            new_text_input = content
-
-        return new_text_input
-
-    def _add_message_to_json(self, openai_json: Dict, message: Optional[Dict]) -> Dict:
-        if message:
-            openai_json["payload"][0]["messages"].append(message)
-
-        return openai_json
-
-    def _add_new_text_input_to_json(
-        self, pa_json: Dict, index: int, new_text_input: str
-    ) -> Dict:
-        if new_text_input:
-            if pa_json["data"][index]["text_input"][0]:
-                pa_json["data"][index]["text_input"][0] = (
-                    pa_json["data"][index]["text_input"][0] + f" {new_text_input}"
-                )
-            else:
-                pa_json["data"][index]["text_input"][0] = new_text_input
-
-        return pa_json
-
-    def _add_new_prompt_to_json(
-        self,
-        pa_json: Dict,
-        index: int,
-        new_prompt: str,
-    ) -> Dict:
-        if new_prompt:
-            if pa_json["data"][index]["payload"][0]["prompt"]:
-                pa_json["data"][index]["payload"][0]["prompt"] += f" {new_prompt}"
-            else:
-                pa_json["data"][index]["payload"][0]["prompt"] = new_prompt
-
-        return pa_json
-
-    def _add_optional_tags_to_openai_json(
-        self,
-        openai_json: Dict,
-        model_name: str = "",
-    ) -> None:
-        payload = openai_json["payload"][0]
-        payload["model"] = model_name
-        if self.config.add_stream:
-            payload["stream"] = True
-        if self.config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN:
-            payload["max_tokens"] = int(
-                random.gauss(
-                    self.config.output_tokens_mean, self.config.output_tokens_stddev
-                )
-            )
-        for key, value in self.config.extra_inputs.items():
-            payload[key] = value
-
-    def _add_optional_tags_to_vllm_json(
-        self,
-        pa_json: Dict,
-        index: int,
-        model_name: str = "",
-    ) -> Dict:
-        row = pa_json["data"][index]
-        row["model"] = model_name
-        if self.config.add_stream:
-            row["stream"] = [True]
-        if self.config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN:
-            number_of_tokens = str(
-                int(
-                    max(
-                        0,
-                        random.gauss(
-                            self.config.output_tokens_mean,
-                            self.config.output_tokens_stddev,
-                        ),
-                    )
-                )
-            )
-            sampling_parameters = {
-                "max_tokens": number_of_tokens,
-            }
-            if self.config.output_tokens_deterministic:
-                sampling_parameters["min_tokens"] = number_of_tokens
-            sampling_parameters_str = json.dumps(sampling_parameters)
-            row["sampling_parameters"] = [sampling_parameters_str]
-        for key, value in self.config.extra_inputs.items():
-            row[key] = [value]
-        if "exclude_input_in_output" not in row:
-            row["exclude_input_in_output"] = [True]
-
-        return pa_json
-
-    def _add_optional_tags_to_trtllm_json(
-        self,
-        pa_json: Dict,
-        index: int,
-        model_name: str = "",
-    ) -> Dict:
-        row = pa_json["data"][index]
-        row["model"] = model_name
-        if self.config.add_stream:
-            row["stream"] = [True]
-        if self.config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN:
-            number_of_tokens = int(
-                random.gauss(
-                    self.config.output_tokens_mean, self.config.output_tokens_stddev
-                )
-            )
-            if self.config.output_tokens_deterministic:
-                row["min_length"] = [number_of_tokens]
-            row["max_tokens"] = [number_of_tokens]
-        for key, value in self.config.extra_inputs.items():
-            row[key] = [value]
-
-        return pa_json
-
-    def _add_optional_tags_to_trtllm_engine_json(
-        self, pa_json: Dict, index: int
-    ) -> Dict:
-        row = pa_json["data"][index]
-        if self.config.add_stream:
-            row["streaming"] = [True]
-        if self.config.output_tokens_mean != DEFAULT_OUTPUT_TOKENS_MEAN:
-            num_tokens = int(
-                random.gauss(
-                    self.config.output_tokens_mean, self.config.output_tokens_stddev
-                )
-            )
-            row["request_output_len"] = [num_tokens]
-            if self.config.output_tokens_deterministic:
-                row["min_length"] = [num_tokens]
-
-        for key, value in self.config.extra_inputs.items():
-            row[key] = [value]
-
-        return pa_json
-
-    def _add_required_tags_to_trtllm_json(
-        self,
-        pa_json: Dict,
-        index: int,
-        default_max_tokens: bool,
-    ) -> Dict:
-        row = pa_json["data"][index]
-        if default_max_tokens:
-            row["max_tokens"] = [DEFAULT_TENSORRTLLM_MAX_TOKENS]
-
-        return pa_json
 
     def _check_for_dataset_name_if_input_type_is_url(self) -> None:
         if (
